** See Ximian Bug #70323.
2005-01-08  Not Zed  <NotZed@Ximian.com>

        ** See Ximian Bug #70323.

        * xntlm-md4.c, xntlm-des.c: convert all unsigned long/long to
        guint32, it expects them to be 32 bits.

+++ xntlm/xntlm-des.h
@@ -4,7 +4,9 @@
 #ifndef _XNTLM_DES_H
 #define _XNTLM_DES_H
 
-typedef unsigned long XNTLM_DES_KS[16][2];
+#include <glib/gtypes.h>
+
+typedef guint32 XNTLM_DES_KS[16][2];
 
 enum {
 	XNTLM_DES_ENCRYPT = 0,
